# Fix sparse_hashed dict performance with sequential keys (wrong hash function)