@alexey-milovidov 2021-12-11
@azat Performance test showed significant performance degradation: https://clickhouse-test-reports.s3.yandex.net/32536/6d31a389f12837f3b8168f17217dadab4f4f4974/performance_comparison/report.html#fail1