@alexey-milovidov 2021-12-11
PS. For our own HashTables, we should use crc32c.