@alexey-milovidov 2021-12-11
We should completely revert #27152 and remove useless files, because we don't use "Arcadia" build system anymore.