@azat 2021-12-11
>@azat Performance test showed significant performance degradation: https://clickhouse-test-reports.s3.yandex.net/32536/6d31a389f12837f3b8168f17217dadab4f4f4974/performance_comparison/report.html#fail1

Yes, the problem is that for sequential keys DefaultHash does not works great, but it works significantly better if those sequential keys are randomly accessed/inserted, I'm going to update performance test for now (to fix the original issue) and get back to this later.

>We should completely revert #27152 and remove useless files, because we don't use "Arcadia" build system anymore.

- #32580 - BTW there is not a lot had been left