@azat 2021-12-13
So now [perf tests](https://clickhouse-test-reports.s3.yandex.net/32536/f4fdea7f8213e2a1ef4611e73ea580d2e5c99aa4/performance_comparison/all-queries.html#all-query-times.hashed_dictionary.0) shows improvement, but the query itself is too slow (there is lots of heuristics in report.py, that does not allow the query to run >~2 seconds each):

Old, s | New, s | Ratio of speedup (-) or slowdown (+) | Relative difference (new − old) / old | p < 0.01 threshold | Test | # | Query
-- | -- | -- | -- | -- | -- | -- | --
12.281 | 5.111 | -2.402x | -0.584 | 0.578 | hashed_dictionary | 1 | SYSTEM RELOAD DICTIONARY SPARSE_HASHED_dictionary

Will rework the perf test.