@alexey-milovidov 2021-12-14
Could you please remove the `SparseHashMap.h` file?
It was added for "arcadia" and we don't use arcadia anymore.

Also need to figure out what is wrong with integration tests...