@azat 2021-12-14
https://github.com/ClickHouse/ClickHouse/runs/4516380764?check_suite_focus=true

```
  Error: fatal: Unable to create '/home/ubuntu/actions-runner/_work/ClickHouse/ClickHouse/.git/modules/contrib/aws/index.lock': File exists.
  
  Another git process seems to be running in this repository, e.g.
  an editor opened by 'git commit'. Please make sure all processes
  are terminated then try again. If it still fails, a git process
  may have crashed in this repository earlier:
  remove the file manually to continue.
...
  Unable to checkout '00b03604543367d7e310cb0993973fdcb723ea79' in submodule path 'contrib/aws'
  Error: The process '/usr/bin/git' failed with exit code 1
```

Something left from the previous run in this workspace?